Don't call producer in go func
If the consumer calls the producer in a go func, we can run into an
issue where we are consuming message faster then we can produce, which
could cause to many go funcs to get created.  After discussing with
Tyler and Will, we decided to make it a blocking call, so we can only
consume as fast as we can produce.  If we need more speed, we can always
scale the event-router out.

Signed-off-by: Ken Herner <ken@astronomer.io>