event-router will retry connecting to Houston is connection fails
Event-router should not lose messages when it loses connection to
houston now.  Instead if it can't connect to houston, it will
rebroadcast the message to the main topic, then stop consuming.  The
consumer will then sleep 5 seconds befor trying to get an auth token
again and start consuming.  If it can't get an auth token, it will sleep
for 5 seconds and continue.  Like a health check.

Signed-off-by: Ken Herner <ken@astronomer.io>